Add files via upload
### Introduction


A heart attack (also known as a myocardial infarction or MI) occurs when a blood clot blocks a coronary artery, causing heart muscle damage and death. Coronary arteries are blood channels that give blood and oxygen to the heart muscle. A coronary artery blockage deprives the heart muscle of blood and oxygen, resulting in heart muscle damage. Chest pain and a feeling of pressure in the chest are symptoms of a heart muscle injury. If blood supply to the heart muscle is not restored within 20 to 40 minutes, permanent heart muscle death will commence. Muscle continues to die for another six to eight hours after which the heart attack is considered "complete."

[https://www.medicinenet.com/heart_attack/article.htm](http://)

### Problem Statement

* Predicting heart attack with some vital statistics of Patients. 

https://www.kaggle.com/rashikrahmanpritom/heart-attack-analysis-prediction-dataset